add specific windows
