+++ pkg/integrations/windows_exporter/config.go
@@ -5,7 +5,6 @@ import (
 	"github.com/grafana/agent/pkg/integrations"
 	integrations_v2 "github.com/grafana/agent/pkg/integrations/v2"
 	"github.com/grafana/agent/pkg/integrations/v2/metricsutils"
-	"github.com/prometheus-community/windows_exporter/pkg/collector"
 )
 
 // DefaultConfig holds the default settings for the windows_exporter integration.
@@ -118,43 +117,6 @@ func (c *Config) InstanceKey(agentKey string) (string, error) {
 // NewIntegration creates an integration based on the given configuration
 func (c *Config) NewIntegration(l log.Logger) (integrations.Integration, error) {
 	return New(l, c)
-}
-
-func (c *Config) ToWindowsExporterConfig() collector.Config {
-	cfg := collector.ConfigDefaults
-	cfg.Dfsr.DfsrEnabledCollectors = c.Dfsr.SourcesEnabled
-	cfg.Exchange.CollectorsEnabled = c.Exchange.EnabledList
-
-	cfg.Iis.SiteInclude = coalesceString(c.IIS.SiteInclude, c.IIS.SiteWhiteList)
-	cfg.Iis.SiteExclude = coalesceString(c.IIS.SiteExclude, c.IIS.SiteBlackList)
-	cfg.Iis.AppInclude = coalesceString(c.IIS.AppInclude, c.IIS.AppWhiteList)
-	cfg.Iis.AppExclude = coalesceString(c.IIS.AppExclude, c.IIS.AppBlackList)
-
-	cfg.Service.ServiceWhereClause = c.Service.Where
-	cfg.Service.UseAPI = c.Service.UseApi == "true"
-
-	cfg.Smtp.ServerInclude = coalesceString(c.SMTP.Include, c.SMTP.WhiteList)
-	cfg.Smtp.ServerExclude = coalesceString(c.SMTP.Exclude, c.SMTP.BlackList)
-
-	cfg.Textfile.TextFileDirectories = c.TextFile.TextFileDirectory
-
-	cfg.Process.ProcessExclude = coalesceString(c.Process.Exclude, c.Process.BlackList)
-	cfg.Process.ProcessInclude = coalesceString(c.Process.Include, c.Process.WhiteList)
-
-	cfg.Net.NicExclude = coalesceString(c.Network.Exclude, c.Network.BlackList)
-	cfg.Net.NicInclude = coalesceString(c.Network.Include, c.Network.WhiteList)
-
-	cfg.Mssql.EnabledCollectors = c.MSSQL.EnabledClasses
-
-	cfg.Msmq.QueryWhereClause = c.MSMQ.Where
-
-	cfg.LogicalDisk.VolumeInclude = coalesceString(c.LogicalDisk.Include, c.LogicalDisk.WhiteList)
-	cfg.LogicalDisk.VolumeExclude = coalesceString(c.LogicalDisk.Exclude, c.LogicalDisk.BlackList)
-
-	cfg.ScheduledTask.TaskInclude = c.ScheduledTask.Include
-	cfg.ScheduledTask.TaskExclude = c.ScheduledTask.Exclude
-
-	return cfg
 }
 
 func coalesceString(v ...string) string {
